+++ src/Entidades/Laboratorio.java
@@ -4,14 +4,14 @@ package Entidades;
 public class Laboratorio {
     
     private int idLaboratorio;
-    private int cuit;
+    private long cuit;
     private String nomLaboratorio;
     private String pais;
     private String domComercial;
     private boolean estado;
     
     // CONSTRUCTOR COMPLETO
-    public Laboratorio(int idLaboratorio, int cuit, String nomLaboratorio, String pais, String domComercial, boolean estado) {
+    public Laboratorio(int idLaboratorio, long cuit, String nomLaboratorio, String pais, String domComercial, boolean estado) {
         this.idLaboratorio = idLaboratorio;
         this.cuit = cuit;
         this.nomLaboratorio = nomLaboratorio;
@@ -21,7 +21,7 @@ public class Laboratorio {
     }
     
     // CONSTRUCTOR SIN ID
-    public Laboratorio(int cuit, String nomLaboratorio, String pais, String domComercial) {
+    public Laboratorio(long cuit, String nomLaboratorio, String pais, String domComercial) {
         this.cuit = cuit;
         this.nomLaboratorio = nomLaboratorio;
         this.pais = pais;
@@ -41,11 +41,11 @@ public class Laboratorio {
         this.idLaboratorio = idLaboratorio;
     }
 
-    public int getCuit() {
+    public long getCuit() {
         return cuit;
     }
 
-    public void setCuit(int cuit) {
+    public void setCuit(long cuit) {
         this.cuit = cuit;
     }
